fix(makefile): use printf instead of echo to print formated messages
Some non standard POSIX shells require an extra argument (-e) to interpret backslash
escape sequences when using `echo` (as it's a built-in command), leading them to not
format the message without it. Standard shells, on the other hand, ignore this
argument and interpret it as another message to print. `printf` always interprets
these escape sequences, which should fix the formatting issues